Refactor AddBountyButton
