+++ packages/ui/src/app/pages/Bounty/components/BountiesHeader.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { PageHeader } from '@/app/components/PageHeader'
-import { AddBountyButton } from '@/bounty/components/AddBountyButton/AddBountyButton'
+import { AddBountyButton } from '@/bounty/components/modalsButtons/AddBountyButton'
 
 import { BountiesTabs } from './BountiesTabs'
