+++ packages/ui/src/bounty/components/modalsButtons/AddBountyButton.tsx
@@ -4,8 +4,10 @@ import { AddBountyModalCall } from '@/bounty/modals/AddBountyModal'
 import { TransactionButton } from '@/common/components/buttons/TransactionButton'
 import { PlusIcon } from '@/common/components/icons/PlusIcon'
 import { useModal } from '@/common/hooks/useModal'
+import { useTranslation } from 'react-i18next'
 
 export const AddBountyButton = () => {
+  const { t } = useTranslation('bounty')
   const { showModal } = useModal()
   const addBountyModal = useCallback(() => {
     showModal<AddBountyModalCall>({
@@ -16,7 +18,7 @@ export const AddBountyButton = () => {
   return (
     <TransactionButton style="primary" size="medium" onClick={addBountyModal}>
       <PlusIcon />
-      Create Bounty
+      {t('buttons.createBounty')}
     </TransactionButton>
   )
 }
