+++ packages/ui/src/services/i18n/dict/en/bounty.json
@@ -244,7 +244,7 @@
     "announceEntry": "Announce Entry",
     "submitWork": "Submit Work",
     "submitJudgement": "Submit Judgement",
-    "cancelBounty": "Cancel Bounty"
+    "createBounty": "Create Bounty"
   },
   "information": {
     "bountyFailed": {
